Performance Adjustments
Removed values that adjust the Global Signed Distance Fields. Though it improves lumen stability, it's quite costly at around 3ms, not to mention that the main cvar required: `r.LumenScene.GlobalSDF.Resolution` is not scalable (it needs to be set in Engine.ini).

Another really expensive option is setting the DiffuseIntegralMethod to use Importance Sample BRDF, combined with a higher TracingOctahedronResolution. It really doesn't change the visuals that much, but since all cvars are scalable, I'm leaving them in @Cine quality. I did remove ShortRangeAO.ApplyDuringIntegration as it's just way too noisy to be usable.

I've also disabled the ability to disable Lumen GI altogether as the game assets simply don't support it. In particular, indoor areas like the Manor can look EXTREMELY bright, which requires adjustments to SkylightIntensity. The problem is when setting this to a lower value, it makes outside areas look way too dim. This means for Lumen off to be viable, the developers themselves need to make this SkylightIntensity change depending on the location.

Lastly, I disabled `r.Shadow.Virtual.Cache.ForceInvalidateDirectional` as it eats up a lot of performance. I'll leave it up to you if you think fixing the occasional pixelated VSM is worth about 5-10fps.